LRDOCS-8961 Extract message listening and destination configuration to new classes.
