+++ docs/dxp/latest/en/developing-applications/core-frameworks/message-bus/using-default-synchronous-messaging/resources/liferay-m4q7.zip/m4q7-baker-impl/src/main/java/com/acme/m4q7/baker/internal/osgi/commands/M4Q7MessageSenderOSGiCommands.java
@@ -1,0 +1,45 @@
+package com.acme.m4q7.baker.internal.osgi.commands;
+
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.messaging.Message;
+import com.liferay.portal.kernel.messaging.MessageBusException;
+import com.liferay.portal.kernel.messaging.sender.SynchronousMessageSender;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+@Component(
+	property = {
+		"osgi.command.function=sendMessage", "osgi.command.scope=m4q7.baker"
+	},
+	service = M4Q7MessageSenderOSGiCommands.class
+)
+public class M4Q7MessageSenderOSGiCommands {
+
+	public void sendMessage(String payload) {
+		try {
+			Message message = new Message();
+
+			message.setPayload(payload);
+			message.setResponseDestinationName("acme/m4q7_baker");
+
+			Object response = _synchronousMessageSender.send(
+				"acme/m4q7_able", message, 10000);
+
+			if (_log.isInfoEnabled()) {
+				_log.info("Response: " + response);
+			}
+		}
+		catch (MessageBusException messageBusException) {
+			_log.error(messageBusException, messageBusException);
+		}
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		M4Q7MessageSenderOSGiCommands.class);
+
+	@Reference(target = "(mode=DEFAULT)")
+	private SynchronousMessageSender _synchronousMessageSender;
+
+}
